[FLINK-12827][table-planner] Extract Optimizer from TableEnvImpl
