+++ flink-table/flink-table-planner/src/test/scala/org/apache/flink/table/plan/RetractionRulesTest.scala
@@ -506,7 +506,8 @@ class StreamTableTestForRetractionUtil extends StreamTableTestUtil {
   }
 
   def verifyTableTrait(resultTable: Table, expected: String): Unit = {
-    val optimized = tableEnv.optimize(toRelNode(resultTable), updatesAsRetraction = false)
+    val optimized = tableEnv.optimizer
+      .optimize(toRelNode(resultTable), updatesAsRetraction = false, tableEnv.getRelBuilder)
     val actual = TraitUtil.toString(optimized)
     assertEquals(
       expected.split("\n").map(_.trim).mkString("\n"),
